more porgress

+++ game.py
@@ -7,17 +7,16 @@
 # temporal_alignment
 # spatial alignment
 
-from curses.panel import new_panel
 import os
 import shutil
 import json
 
-from numpy import full
-
 from video import Video
 from data import Data
+from utilities.files import File
 from utilities.timestamp_extraction import extract_timestamps_plus_trim
 from utilities.timestamp_post_processing import postprocess_timestamps
+from utilities.timestamp_visualization import viz_extraction
 
 EXAMPLE_PATH = "/content/drive/MyDrive/Research/sportvu-plus-videos/statvu-plus-plus/01.14.2016.LAL.at.GSW.TNT/01.14.2016.LAL.at.GSW.TNT.mp4"
 
@@ -30,11 +29,12 @@ class Game:
         self.video = video
         self.title = video.title
         self.network = video.network
+        self.folder_path = f"videos-plus-data/{video.title}.{video.network}"
 
     @classmethod
     def __init_with_title__(cls, title: str, network: str) -> object:
         full_title = f"{title}.{network}"
-        path_to_data = f"statvu-raw-data/{full_title}.7z"
+        path_to_data = f"statvu-raw-data/{title}.7z"
         path_to_video = f"unprocessed-videos/{full_title}.mp4"
         return Game(Data(path_to_data), Video(path_to_video))
 
@@ -42,8 +42,7 @@ class Game:
         """Perform temporal and spatial alignment for an NBA basketball game."""
 
         self.create_game_folder()
-        normalized_video_path = self.video.normalize()
-        self.replace_path(self.video.path, normalized_video_path)
+        self.normalize_video()
         self.temporal_alignment()
         self.spatial_alignment()
 
@@ -58,47 +57,71 @@ class Game:
                 print(f"Error creating folder with path {new_folder_path}.")
                 raise Exception
         else:
-            print(f"Error: folder already exists ar {new_folder_path}.")
+            print(f"Folder already exists at {new_folder_path}.")
 
-        # copy data to new dir
-        if not self.data.is_zipped:
-            self.data.unzip(to=new_folder_path)
-        else:
-            new_data_path = f"{new_folder_path}/{self.title}.{self.network}.json"
-            shutil.copy(self.data.path, new_data_path)
+        # copy data to new dir, if data is zipped: unzip
+        new_data_path = f"{new_folder_path}/{self.title}.{self.network}.json"
+
+        if not os.path.exists(new_data_path):
+            if self.data.is_zipped:
+                temp_data_path = self.data.unzip(to=new_folder_path)
+                os.rename(temp_data_path, new_data_path)
+            else:
+                shutil.copy(self.data.path, new_data_path)
             self.data.path = os.path.abspath(new_data_path)
+            assert self.data.path == os.path.abspath(new_data_path)
+            assert os.path.exists(new_data_path)
+        else:
+            print(f"Data file exists and formatted at: {new_data_path}.")
 
         # copy video to new dir
         new_video_path = f"{new_folder_path}/{self.title}.{self.network}.mp4"
-        shutil.copy(self.video.path, new_video_path)
-        self.video.path = os.path.abspath(new_video_path)
+        if not os.path.exists(new_video_path):
+            shutil.copy(self.video.path, new_video_path)
+            self.video.path = os.path.abspath(new_video_path)
+            assert os.path.exists(new_video_path)
+        else:
+            print(
+                f"Video already exists at {new_video_path}. Will normalize if necessary.")
+            self.video.path = new_video_path
+        assert self.video.path == new_video_path
+
+    def normalize_video(self) -> None:
+        """Normalize a video to dim 1280x720 w/ 25 FPS."""
+
+        if not self.video.is_normalized():
+            normalized_video_path = self.video.normalize()
+            File.replace_path(self.video.path, normalized_video_path)
+            assert os.path.exists(normalized_video_path)
 
     def temporal_alignment(self) -> None:
         """Extract video timestamps and add to statvu data."""
 
         timestamps_path = f"videos-plus-data/{self.title}.{self.network}/timestamps.json"
+        try:
+            assert not os.path.exists(timestamps_path)
+        except:
+            print("Error: extracted timestamps already exist. Please remove.")
+            raise Exception
         timestamps = json.dumps(extract_timestamps_plus_trim(
             self.video.path, self.network), indent=4)
+        # video exists after being trimmed + replaced
+        assert os.path.exists(self.video.path)
         try:
             with open(timestamps_path, "w") as outfile:
                 outfile.write(timestamps)
         except:
             print(
-                f"Error: could now save extracted timestamps to {timestamps_path}.")
+                f"Error: could not save extracted timestamps to {timestamps_path}.")
             raise Exception
         modified_timestamps_path = postprocess_timestamps(timestamps_path)
 
-        self.replace_path(timestamps_path, modified_timestamps_path)
+        raise Exception  # break
+        File.replace_path(timestamps_path, modified_timestamps_path)
 
     def spatial_alignment(self) -> None:
         pass
 
-    @classmethod
-    def replace_path(cls, old_path: str, new_path: str) -> None:
-        """Replace the file at old_path with file at new path. Rename new_path to old_path."""
-
-        try:
-            os.remove(old_path)
-            os.rename(new_path, old_path)
-        except:
-            print(f"Error attemping to replace {old_path} with {new_path}.")
+    def visualize_timestamp_extraction(self):
+        viz_extraction(self.video.path, self.data.path, f"")
+        pass
